♻️ リファクタリング
